EDGE-732: add optional deviceApiVersion field to WebRTC createParticipant call.
This commit adds an optional "deviceApiVersion" parameter to the WebRTC "Participant" schema which will allow creating participants under either "v2" or "v3" of the device API. Once this is released and accompanying messaging is sent out, customers will have 90 days to manually test or transition to "v3" of the device API using this field if they wish to; at the end of that period we will change the default to "v3" and sunset "v2" as soon after as possible.

In addition to being needed for internal reasons, this change is necessitated by the pending deprecation of WebRTC "Plan B" in favor of "Unified Plan" in Google Chrome in August 2021.